char encodings now use UByte type instead of short

+++ codeGeneration/test/helpers/Dummies.kt
@@ -27,11 +27,11 @@ internal val DummyMemsizer = object : IMemSizer {
 }
 
 internal val DummyStringEncoder = object : IStringEncoding {
-    override fun encodeString(str: String, altEncoding: Boolean): List<Short> {
+    override fun encodeString(str: String, altEncoding: Boolean): List<UByte> {
         return emptyList()
     }
 
-    override fun decodeString(bytes: List<Short>, altEncoding: Boolean): String {
+    override fun decodeString(bytes: List<UByte>, altEncoding: Boolean): String {
         return ""
     }
 }
